+++ src/xtd_drawing_native/include/xtd/drawing/native/image.hpp
@@ -27,14 +27,14 @@ namespace xtd {
         static size_t flags(intptr_t image);
         static size_t frame_resolutions(intptr_t image);
         static float horizontal_resolution(intptr_t image);
-        static void physical_dimension(intptr_t image, size_t& width, size_t& height);
+        static void physical_dimension(intptr_t image, int32_t& width, int32_t& height);
         static std::vector<int32_t> property_id_list(intptr_t image);
         static std::vector<property_item> property_items(intptr_t image);
         static size_t pixel_format(intptr_t image);
         static size_t raw_format(intptr_t image);
-        static void size(intptr_t image, size_t& width, size_t& height);
+        static void size(intptr_t image, int32_t& width, int32_t& height);
         static float vertical_resolution(intptr_t image);
-        static void get_pixel(intptr_t image, size_t x, size_t y, argb& color);
+        static void get_pixel(intptr_t image, int32_t x, int32_t y, argb& color);
       };
     }
   }
